Fix projectile entities.
