+++ src/main/java/cn/nukkit/entity/impl/projectile/EntityFireball.java
@@ -1,12 +1,11 @@
 package cn.nukkit.entity.impl.projectile;
 
 import cn.nukkit.entity.EntityType;
-import cn.nukkit.entity.impl.BaseEntity;
 import cn.nukkit.entity.projectile.Fireball;
 import cn.nukkit.level.chunk.Chunk;
 import cn.nukkit.nbt.tag.CompoundTag;
 
-public class EntityFireball extends BaseEntity implements Fireball {
+public class EntityFireball extends EntityProjectile implements Fireball {
     public EntityFireball(EntityType<?> type, Chunk chunk, CompoundTag nbt) {
         super(type, chunk, nbt);
     }
