+++ src/main/java/cn/nukkit/entity/impl/projectile/EntityLlamaSpit.java
@@ -1,12 +1,11 @@
 package cn.nukkit.entity.impl.projectile;
 
 import cn.nukkit.entity.EntityType;
-import cn.nukkit.entity.impl.BaseEntity;
 import cn.nukkit.entity.projectile.LlamaSpit;
 import cn.nukkit.level.chunk.Chunk;
 import cn.nukkit.nbt.tag.CompoundTag;
 
-public class EntityLlamaSpit extends BaseEntity implements LlamaSpit {
+public class EntityLlamaSpit extends EntityProjectile implements LlamaSpit {
     public EntityLlamaSpit(EntityType<?> type, Chunk chunk, CompoundTag nbt) {
         super(type, chunk, nbt);
     }
