+++ src/main/java/cn/nukkit/entity/impl/projectile/EntityShulkerBullet.java
@@ -1,12 +1,11 @@
 package cn.nukkit.entity.impl.projectile;
 
 import cn.nukkit.entity.EntityType;
-import cn.nukkit.entity.impl.BaseEntity;
 import cn.nukkit.entity.projectile.ShulkerBullet;
 import cn.nukkit.level.chunk.Chunk;
 import cn.nukkit.nbt.tag.CompoundTag;
 
-public class EntityShulkerBullet extends BaseEntity implements ShulkerBullet {
+public class EntityShulkerBullet extends EntityProjectile implements ShulkerBullet {
     public EntityShulkerBullet(EntityType<?> type, Chunk chunk, CompoundTag nbt) {
         super(type, chunk, nbt);
     }
