+++ src/main/java/cn/nukkit/entity/impl/projectile/EntitySmallFireball.java
@@ -1,12 +1,11 @@
 package cn.nukkit.entity.impl.projectile;
 
 import cn.nukkit.entity.EntityType;
-import cn.nukkit.entity.impl.BaseEntity;
 import cn.nukkit.entity.projectile.SmallFireball;
 import cn.nukkit.level.chunk.Chunk;
 import cn.nukkit.nbt.tag.CompoundTag;
 
-public class EntitySmallFireball extends BaseEntity implements SmallFireball {
+public class EntitySmallFireball extends EntityProjectile implements SmallFireball {
     public EntitySmallFireball(EntityType<?> type, Chunk chunk, CompoundTag nbt) {
         super(type, chunk, nbt);
     }
